Combined Nik and I's code
Had to manually do source control on what Nik uploaded last week. If we worked on the same file, I had to manually update it by hand. Because Nik does not have my code, so if we both worked on something, I wouldn't have had Nik's code, and vice versa, so I had to manually make sure that all of the functionality was preserved. This was extremely time consuming.